requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.3.2
beanie==1.25.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cloudinary==1.39.0
httpx==0.27.0
websockets==12.0
python-dotenv==1.0.1
hashlib2==1.0.0
pytest==8.1.1
pytest-asyncio==0.23.6
aiofiles==23.2.1
Pillow==10.3.0
